Add Reason based initial UI
This commit adds the configuration for melange compilation, along with
the actual frontend source code.

There's no real sharing between the frontend and the server code, as of
now, since I haven't yet figured out how to do sharing and builds in
different modes with different preprocessors.

Use marked to convert Markdown to HTML in the Reason UI